+++ spec/rubocop/cop/style/nil_comparison_spec.rb
@@ -17,12 +17,6 @@ describe Rubocop::Cop::Style::NilComparison do
     expect(cop.offenses.size).to eq(1)
   end
 
-  it 'registers an offense for === nil' do
-    inspect_source(cop,
-                   ['x != nil'])
-    expect(cop.offenses.size).to eq(1)
-  end
-
   it 'works with lambda.()' do
     inspect_source(cop, ['a.(x) == nil'])
     expect(cop.offenses.size).to eq(1)